@@ -4,64 +4,76 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~14.2.0",
-    "@angular/common": "~14.2.0",
-    "@angular/compiler": "~14.2.0",
-    "@angular/core": "~14.2.0",
-    "@angular/forms": "~14.2.0",
-    "@angular/platform-browser": "~14.2.0",
-    "@angular/platform-browser-dynamic": "~14.2.0",
-    "@angular/router": "~14.2.0",
-    "@nestjs/common": "^9.0.0",
-    "@nestjs/core": "^9.0.0",
-    "@nestjs/platform-express": "^9.0.0",
-    "@nrwl/angular": "14.7.6",
+    "@angular/animations": "16.2.1",
+    "@angular/common": "16.2.1",
+    "@angular/compiler": "16.2.1",
+    "@angular/core": "16.2.1",
+    "@angular/forms": "16.2.1",
+    "@angular/platform-browser": "16.2.1",
+    "@angular/platform-browser-dynamic": "16.2.1",
+    "@angular/router": "16.2.1",
+    "@nestjs/common": "10.0.2",
+    "@nestjs/core": "10.0.2",
+    "@nestjs/platform-express": "10.0.2",
+    "@ngrx/component-store": "~16.0.0",
+    "@ngrx/effects": "~16.0.0",
+    "@ngrx/entity": "~16.0.0",
+    "@ngrx/router-store": "~16.0.0",
+    "@ngrx/store": "^16.2.0",
+    "@nx/angular": "16.7.2",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.5.0",
+    "rxjs": "7.8.1",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.2.0",
-    "@angular-eslint/eslint-plugin": "~14.0.4",
-    "@angular-eslint/eslint-plugin-template": "~14.0.4",
-    "@angular-eslint/template-parser": "~14.0.4",
-    "@angular/cli": "~14.2.0",
-    "@angular/compiler-cli": "~14.2.0",
-    "@angular/language-service": "~14.2.0",
-    "@nestjs/schematics": "^9.0.0",
-    "@nestjs/testing": "^9.0.0",
-    "@nrwl/cli": "14.7.6",
-    "@nrwl/cypress": "14.7.6",
-    "@nrwl/eslint-plugin-nx": "14.7.6",
-    "@nrwl/jest": "14.7.6",
-    "@nrwl/linter": "14.7.6",
-    "@nrwl/nest": "14.7.6",
-    "@nrwl/node": "14.7.6",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/workspace": "14.7.6",
-    "@types/jest": "28.1.1",
+    "@angular-devkit/build-angular": "16.2.0",
+    "@angular-devkit/core": "16.2.0",
+    "@angular-devkit/schematics": "16.2.0",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
+    "@angular/cli": "~16.2.0",
+    "@angular/compiler-cli": "16.2.1",
+    "@angular/language-service": "16.2.1",
+    "@nestjs/schematics": "10.0.1",
+    "@nestjs/testing": "10.0.2",
+    "@ngrx/schematics": "~16.0.0",
+    "@nx/cypress": "16.7.2",
+    "@nx/eslint-plugin": "16.7.2",
+    "@nx/jest": "16.7.2",
+    "@nx/linter": "16.7.2",
+    "@nx/nest": "16.7.2",
+    "@nx/node": "16.7.2",
+    "@nx/workspace": "16.7.2",
+    "@schematics/angular": "16.2.0",
+    "@types/jest": "29.4.4",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.36.1",
-    "@typescript-eslint/parser": "^5.36.1",
-    "cypress": "^10.7.0",
-    "eslint": "~8.15.0",
+    "@typescript-eslint/eslint-plugin": "5.60.1",
+    "@typescript-eslint/parser": "5.60.1",
+    "autoprefixer": "^10.4.0",
+    "cypress": "^12.16.0",
+    "eslint": "8.46.0",
     "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "jest": "28.1.1",
-    "jest-environment-jsdom": "28.1.1",
-    "jest-preset-angular": "~12.2.2",
-    "nx": "14.7.6",
+    "eslint-plugin-cypress": "2.13.4",
+    "jasmine-marbles": "~0.9.1",
+    "jest": "29.4.3",
+    "jest-environment-jsdom": "29.4.3",
+    "jest-preset-angular": "13.1.1",
+    "nx": "16.7.2",
+    "nx-cloud": "16.4.0-beta.1",
+    "postcss": "^8.4.5",
     "prettier": "^2.6.2",
-    "ts-jest": "28.0.5",
+    "tailwindcss": "^3.0.2",
+    "ts-jest": "29.1.1",
     "ts-node": "10.9.1",
-    "typescript": "~4.8.2"
+    "typescript": "5.1.6"
   }
 }
