@@ -1,4 +1,4 @@
-import { Controller, Get } from '@nestjs/common';
+import { Body, Controller, Get, Post } from '@nestjs/common';
 
 import { Message } from '@ngrx-demo-client/api-interfaces';
 
@@ -11,5 +11,15 @@ export class AppController {
   @Get('hello')
   getData(): Message {
     return this.appService.getData();
+  }
+
+  @Get('tasks')
+  getTasks() {
+    return this.appService.get();
+  }
+
+  @Post('tasks')
+  createTask(@Body() dto) {
+    return this.appService.create(dto);
   }
 }
